// gravitee-apim-portal-webui/projects/portal-webclient-sdk/src/lib/model/referenceMetadataFormatType.ts
/**
 * Gravitee.io Portal Rest API
 * API dedicated to the devportal part of Gravitee
 *
 * Contact: contact@graviteesource.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export type ReferenceMetadataFormatType = 'STRING' | 'NUMERIC' | 'BOOLEAN' | 'DATE' | 'MAIL' | 'URL';

export const ReferenceMetadataFormatType = {
    STRING: 'STRING' as ReferenceMetadataFormatType,
    NUMERIC: 'NUMERIC' as ReferenceMetadataFormatType,
    BOOLEAN: 'BOOLEAN' as ReferenceMetadataFormatType,
    DATE: 'DATE' as ReferenceMetadataFormatType,
    MAIL: 'MAIL' as ReferenceMetadataFormatType,
    URL: 'URL' as ReferenceMetadataFormatType
};